update structure and split logic for partial installs
