+++ src/uQCme/app/__init__.py
@@ -1,0 +1,9 @@
+"""Streamlit app module for uQCme.
+
+Note: This module requires the 'app' or 'all' extras to be installed:
+    pip install uqcme[app]
+"""
+
+from uQCme.app.main import main
+
+__all__ = ["main"]
