+++ src/uQCme/app/main.py
@@ -6,8 +6,8 @@ A Streamlit web application for visualizing and exploring microbial QC results
 generated by the uQCme CLI tool. This module focuses on data tables and
 interactive data exploration, with plotting functionality handled by plot.py.
 
-Note: This module requires the 'full' extras to be installed:
-    pip install uqcme[full]
+Note: This module requires the 'app' or 'all' extras to be installed:
+    pip install uqcme[app]
 """
 
 import os
@@ -18,7 +18,7 @@ import pandas as pd
 import yaml
 from pathlib import Path
 from typing import Dict, Any, Optional
-from uQCme.plot import QCPlotter, get_available_metrics
+from uQCme.app.plot import QCPlotter, get_available_metrics
 from uQCme.core.loader import load_data_from_config, load_config_from_file
 from uQCme.core.engine import QCProcessor
 from uQCme.core.config import UQCMeConfig, DataInput
