+++ src/uQCme/cli/__init__.py
@@ -1,0 +1,5 @@
+"""CLI module for uQCme."""
+
+from uQCme.cli.main import main
+
+__all__ = ["main"]
